pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-tg"
version = "0.6.0"
description = "Claude Code <-> Telegram bridge"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "python-telegram-bot>=21.0",
    "mcp>=1.0",
    "groq>=0.4",
]

[project.scripts]
claude-tg = "claude_tg.__main__:main"
claude-tg-mcp = "claude_tg.mcp_server:main"
